+++ internal/commands/kubernetes/show_test.go
@@ -112,36 +112,13 @@ func TestShowCommand(t *testing.T) {
     Zone:                de-fra1                              
     Operational state:   running                              
 
-  
-  Node group 1 (upcloud-go-sdk-unit-test):
-    Name:                   upcloud-go-sdk-unit-test              
-    Count:                  4                                     
-    Plan:                   2xCPU-4GB                             
-    State:                  running                               
-    Storage UUID:           storage-uuid                          
-    Storage name:           Test storage                          
-    Kubelet args:           somekubeletkey=somekubeletvalue       
-    Labels:                 managedBy=upcloud-go-sdk-unit-test    
-                            another=label-thing                   
-    Taints:                 sometaintkey=sometaintvalue:NoExecute 
-                            sometaintkey=sometaintvalue:NoExecute 
-                            sometaintkey=sometaintvalue:NoExecute 
-    Utility network access: yes                                   
+  Node groups:
 
-  
-  Node group 2 (upcloud-go-sdk-unit-test-2):
-    Name:                   upcloud-go-sdk-unit-test-2               
-    Count:                  8                                        
-    Plan:                   4xCPU-8GB                                
-    State:                  pending                                  
-    Storage UUID:           storage-uuid-2                           
-    Storage name:           Test storage                             
-    Kubelet args:           somekubeletkey2=somekubeletvalue2        
-    Labels:                 managedBy=upcloud-go-sdk-unit-test-2     
-                            another2=label-thing-2                   
-    Taints:                 sometaintkey2=sometaintvalue2:NoSchedule 
-    Utility network access: no                                       
-
+     Name                         Count   Plan        Anti affinity   Utility network access   State   
+    ──────────────────────────── ─────── ─────────── ─────────────── ──────────────────────── ─────────
+     upcloud-go-sdk-unit-test         4   2xCPU-4GB   no              yes                      running 
+     upcloud-go-sdk-unit-test-2       8   4xCPU-8GB   no              no                       pending 
+    
 `
 
 	mService := smock.Service{}